move print id to uid_util.h
