+++ be/src/runtime/tmp_file_mgr.cc
@@ -28,7 +28,7 @@
 // #include <gutil/strings/join.h>
 
 #include "olap/olap_engine.h"
-#include "util/debug_util.h"
+#include "util/uid_util.h"
 #include "util/disk_info.h"
 #include "util/filesystem_util.h"
 #include "runtime/exec_env.h"
